:sparkles: Remove options, use libc settings instead
Build options are not the appropriate choice for this package. Options
for tool packages are only available or controllable via build profiles
or tool requires calls in python. This becomes very annoying because it
causes the developer to do stuff like this to make a build work:

```
conan build . -pr target -pr compiler -pr:b build_profile
```

Where build_profile is defined as:

```
include(default)

[options]
custom_libc=True
```

For context, this injects the default build profile and adds the option.
Target profiles cannot effect build options so it must go in the build
profile. An issue with this is that its easy to forget and fail at
runtime.

Preference is placed on reducing the number of arguments the user has to
write. Already having two profiles to select is annoying enough, but a
3rd profile that causes a silent bug in your code is unacceptable. The
current design is to do the following:

```
conan build . -pr target -pr compiler
```

This uses the new `libc` binary model settings in conan-config
(https://github.com/libhal/conan-config/pull/4) as well as the CMake
toolchain.cmake and the CMakeToolchain generator. Application binaries
that need to use a custom libc must do the following:

1. Add "libc" to the list of settings for the conanfile.py
2. Pass the `libc` variable value to the CMakeToolchain generator like
   so:

```
def generate(self):
  tc = CMakeToolchain(self)
  tc.cache_variables["CONAN_LIBC"] = str(self.settings.libc)
  tc.generate()
```

All other libraries and packages that do not care about using a custom
libc simply omit these and the default libc flags will be injected.