:sparkles: New default options
Add support for the following additional options:

  - default_arch
  - default_specs
  - lto
  - fat_lto
  - function_sections
  - data_sections
  - gc_sections

These options are set to TRUE to provide great defaults that most
projects are going to want to take advantage of. This also allows
developers to control these options from their profiles. The choice to
move away from using `[conf]` within our pre-made profiles is due to
the behavior where it prevents any changes to those `conf` items from
packages. Its a way to hard override the options at the top most level
which is not desirable from our point of view.